+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 ## Unreleased
 
 ### Updated
-1. Bumped Go version to 1.
+1. Bumped Go version to 1.22
+2. Reworked member initialisation to resolve group names against system group names using
+   group ID because Wild Apricot does not keep the member groups fields consistent.
 
 
 ## [0.8.7](https://github.com/uhppoted/uhppoted-app-wild-apricot/releases/tag/v0.8.7) - 2023-12-01